movement is mostly fixed. still some weird behaviors due to rounding

+++ src/components/AI.hpp
@@ -4,13 +4,13 @@
 #include "../ai/AIState.hpp"
 #include "../ai/Blackboard.hpp"
 #include "../behaviortree/BehaviorTree.hpp"
-#include "../engine/types/Vec2f.hpp"
+#include "../engine/types/Vec2i.hpp"
 #include "behaviortree_cpp/bt_factory.h"
 #include <memory>
 #include <vector>
 
 struct AI {
-	Vec2f originalPosition; // The fallback position for AI to return to
+	Vec2i originalPosition; // The fallback position for AI to return to
 
 	// AIRole role = AIRole::Neutral;
 	AIState state = AIState::Unaware;
@@ -21,8 +21,8 @@ struct AI {
 	const float detectionThreshold = 2.0; // Time to transition to searching
 	const float searchTimeout = 5.0;      // Timeout for searching
 
-	Vec2f targetPosition{-1.0f, -1.0f}; // Current target
-	std::vector<Vec2f> path = {}; // Current path to the target
+	Vec2i targetPosition{-1, -1}; // Current target
+	std::vector<Vec2i> path = {}; // Current path to the target
 	int pathIndex = 0;
 
 	// TODO: handle serialization, when the time comes
