+++ src/entities/player.hpp
@@ -11,10 +11,10 @@
 #include "../components/Rotatable.hpp"
 #include "../constants.hpp"
 #include "../ecs/ECSManager.hpp"
-#include "../engine/types/Vec2f.hpp"
+#include "../engine/types/Vec2i.hpp"
 
 // for playerSpriteSheetY use one of the constants PLAYER_{RED|BLUE|WHITE}_SPRITE_SHEET_Y
-Entity instantiatePlayerEntity(ECSManager &ecs, Vec2f positionInTiles, Rotation rotation = SOUTH,
+Entity instantiatePlayerEntity(ECSManager &ecs, Vec2i positionInTiles, Rotation rotation = SOUTH,
                                int playerSpriteSheetY = PLAYER_SPRITE_SHEET_Y_RED)
 {
 	Entity player = ecs.addEntity();
